+++ docs/manual/docs/about/release-notes.md
@@ -99,7 +99,7 @@
 - [graph teams app remove](../cmd/graph/teams/teams-app-remove.md) - removes a Teams app from the organization's app catalog [#825](https://github.com/pnp/office365-cli/issues/825)
 - [graph teams app update](../cmd/graph/teams/teams-app-update.md) - updates Teams app in the organization's app catalog [#827](https://github.com/pnp/office365-cli/issues/827)
 - [graph teams channel list](../cmd/graph/teams/teams-channel-list.md) - lists channels in the specified Microsoft Teams team [#586](https://github.com/pnp/office365-cli/issues/586)
-- [graph teams user remove](../cmd/graph/teams/teams-user-remove.md) - removes the specified user from the specified Microsoft Teams team [#757](https://github.com/pnp/office365-cli/issues/757)
+- [graph teams user remove](../cmd/graph/o365group/o365group-user-remove.md) - removes the specified user from the specified Microsoft Teams team [#757](https://github.com/pnp/office365-cli/issues/757)
 - [graph teams user set](../cmd/graph/teams/teams-user-set.md) - updates role of the specified user in the given Microsoft Teams team [#760](https://github.com/pnp/office365-cli/issues/760)
 
 ### Changes